Load variables before rendering Data Explorer
