+++ ui/src/dataExplorer/components/DataExplorer.tsx
@@ -4,6 +4,7 @@ import {connect} from 'react-redux'
 
 // Components
 import TimeMachine from 'src/timeMachine/components/TimeMachine'
+import GetVariables from 'src/shared/components/GetVariables'
 
 // Actions
 import {setActiveTimeMachine} from 'src/timeMachine/actions'
@@ -32,7 +33,9 @@ class DataExplorer extends PureComponent<DispatchProps, {}> {
     return (
       <div className="data-explorer">
         <HoverTimeProvider>
-          <TimeMachine />
+          <GetVariables>
+            <TimeMachine />
+          </GetVariables>
         </HoverTimeProvider>
       </div>
     )
